+++ wasm/src/org.graalvm.wasm/src/org/graalvm/wasm/WasmTable.java
@@ -65,7 +65,7 @@ public final class WasmTable extends EmbedderDataHolder implements TruffleObject
     /**
      * @see #elemType()
      */
-    private final byte elemType;
+    private final int elemType;
 
     /**
      * @see #minSize()
@@ -86,7 +86,7 @@ public final class WasmTable extends EmbedderDataHolder implements TruffleObject
     private Object[] elements;
 
     @TruffleBoundary
-    private WasmTable(int declaredMinSize, int declaredMaxSize, int initialSize, int maxAllowedSize, byte elemType, Object initialValue) {
+    private WasmTable(int declaredMinSize, int declaredMaxSize, int initialSize, int maxAllowedSize, int elemType, Object initialValue) {
         assert compareUnsigned(declaredMinSize, initialSize) <= 0;
         assert compareUnsigned(initialSize, maxAllowedSize) <= 0;
         assert compareUnsigned(maxAllowedSize, declaredMaxSize) <= 0;
@@ -103,11 +103,11 @@ public final class WasmTable extends EmbedderDataHolder implements TruffleObject
         this.elemType = elemType;
     }
 
-    public WasmTable(int declaredMinSize, int declaredMaxSize, int maxAllowedSize, byte elemType) {
+    public WasmTable(int declaredMinSize, int declaredMaxSize, int maxAllowedSize, int elemType) {
         this(declaredMinSize, declaredMaxSize, declaredMinSize, maxAllowedSize, elemType, WasmConstant.NULL);
     }
 
-    public WasmTable(int declaredMinSize, int declaredMaxSize, int maxAllowedSize, byte elemType, Object initialValue) {
+    public WasmTable(int declaredMinSize, int declaredMaxSize, int maxAllowedSize, int elemType, Object initialValue) {
         this(declaredMinSize, declaredMaxSize, declaredMinSize, maxAllowedSize, elemType, initialValue);
     }
 
@@ -156,9 +156,10 @@ public final class WasmTable extends EmbedderDataHolder implements TruffleObject
      * <p>
      * This table can only be imported with an equivalent elem type.
      *
-     * @return Either {@link WasmType#FUNCREF_TYPE} or {@link WasmType#EXTERNREF_TYPE}.
+     * @return Either {@link WasmType#FUNCREF_TYPE}, {@link WasmType#EXTERNREF_TYPE} or some
+     *         concrete reference type.
      */
-    public byte elemType() {
+    public int elemType() {
         return elemType;
     }
